Add comments and organize backend
