+++ client/src/reducers/index.js
@@ -1,3 +1,5 @@
+// client.src.reducers.index.js
+
 import { combineReducers } from "redux";
 import authReducer from "./authReducer";
 import errorReducer from "./errorReducer";
