+++ config/keys.js
@@ -1,3 +1,5 @@
+// config.keys.js
+
 module.exports = {
   secretOrKey: "secret"
 };
