+++ config/passport.js
@@ -1,3 +1,5 @@
+// config.passport.js
+
 const JwtStrategy = require("passport-jwt").Strategy;
 const ExtractJwt = require("passport-jwt").ExtractJwt;
 const keys = require("./keys");
