+++ controllers/userControls/currentUser.js
@@ -1,0 +1,11 @@
+// controllers.userControls.getCurrentUser.js
+
+function getCurrentUser (req, res) {
+    res.json({
+        id: req.user.id,
+        name: req.user.name,
+        email: req.user.email,
+    });
+}
+
+module.exports = getCurrentUser;
