+++ models/User.js
@@ -1,3 +1,5 @@
+// models.User.js
+
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
