+++ models/index.js
@@ -1,3 +1,5 @@
+// models.index.js
+
 module.exports = {
   Show: require("./Show"),
   User: require("./User")
