+++ routes/api/index.js
@@ -1,7 +1,5 @@
 // routes.api.index.js
 
-// routes.api.index.js
-
 const router = require("express").Router();
 const showRoutes = require("./shows");
 const externalRoutes = require("./external");
