+++ server.js
@@ -1,3 +1,5 @@
+// server.js
+
 require('dotenv').config();
 const express = require("express");
 const logger = require("morgan");
